lib-franklin updated
