+++ blocks/calendar-events/calendar-events.js
@@ -1,4 +1,4 @@
-import { decorateIcons } from '../../scripts/lib-franklin.js';
+import { decorateIcons } from '../../scripts/aem.js';
 
 export default async function decorate(block) {
   block.classList.add(`calendar-events-${block.children.length}`);
