+++ blocks/cards/icon-cta.js
@@ -1,4 +1,4 @@
-import { decorateIcons } from '../../scripts/lib-franklin.js';
+import { decorateIcons } from '../../scripts/aem.js';
 
 /**
  * Builds the Icons variation of the cards block.
