+++ blocks/cards/icon.js
@@ -1,4 +1,4 @@
-import { decorateIcons } from '../../scripts/lib-franklin.js';
+import { decorateIcons } from '../../scripts/aem.js';
 
 export default async function decorate(block) {
   let count;
