+++ tools/importer/forms/ui/index.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta name="viewport" content="width=device-width, initial-scale=1"/>
-        <script src="/scripts/lib-franklin.js" type="module"></script>
+        <script src="/scripts/aem.js" type="module"></script>
         <script src="/scripts/scripts.js" type="module"></script>
         <link rel="stylesheet" href="/styles/styles.css"/>
         <link rel="icon" href="data:,">
